requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pyodbc